fix(stella-plugin): split panel.rs and refuse a bidi-flipped plugin name
Two small, independent defects in stella-plugin, landed together:

- panel.rs sat 11 lines from the gate's 1500-line ceiling. Split it the
  way manifest.rs already splits: panel.rs beside panel/, with the test
  module in panel/tests.rs and the three error enums (PanelRefusal,
  PanelTextError, PanelOverflow) in panel/error.rs, re-exported so every
  public path (stella_plugin::PanelText and its siblings) is unchanged.
  Witness: panel.rs now sits at 1055 lines, well clear of the ceiling
  the gate's check-file-size guard was enforcing before this change.

- validate_plugin_name asked first_control_character but not
  first_bidi_control, so a plugin could ship a name carrying U+202E or
  its eleven siblings — the Trojan Source shape (CVE-2021-42574) —
  printed under Stella's own border, next to the consent prompt a
  person reads to decide whether to trust the plugin. It now asks both,
  returning a new ManifestError::NameCarriesBidiControl naming the
  character and its position, with the joiners U+200C/U+200D still
  allowed. Witness: a_name_that_reorders_itself_with_a_bidi_control_is_refused_by_position
  (crates/stella-plugin/src/drawable.rs) and
  a_plugin_name_that_reorders_itself_with_a_bidi_override_does_not_load
  (crates/stella-plugin/tests/panel_contract.rs) both fail on the old
  code — a manifest naming U+202E loaded clean — and pass on the new.

Tests deleted: None.

Closes #5922
Closes #5917
